requirements.txt
matplotlib==3.6.2
pandas==1.5.2
python-dotenv==0.21.0
requests==2.28.1
yfinance==0.1.95